Remove federated types/fields from sdl render
